--- a/crates/rome_formatter/src/formatter.rs
+++ b/crates/rome_formatter/src/formatter.rs
@@ -201,6 +201,7 @@ impl<'buf, Context> Formatter<'buf, Context> {
         match elements.len() {
             0 => None,
             // Doesn't get cheaper than calling clone, use the element directly
+            // SAFETY: Safe because of the `len == 1` check in the match arm.
             1 => Some(elements.pop().unwrap()),
             _ => Some(FormatElement::Interned(Interned::new(elements))),
         }