--- a/crates/rome_formatter/src/prelude.rs
+++ b/crates/rome_formatter/src/prelude.rs
@@ -10,6 +10,9 @@ pub use crate::trivia::{
 
 pub use crate::verbatim::{format_suppressed_node, format_unknown_node, format_verbatim_node};
 
+pub use crate::format_element::document::Document;
+pub use crate::format_element::signal::{LabelId, Signal, SignalKind};
+
 pub use crate::{
     best_fitting, dbg_write, format, format_args, write, Buffer as _, BufferExtensions, Format,
     Format as _, FormatError, FormatResult, FormatRule, FormatWithRule as _, SimpleFormatContext,