… crates/rome_formatter/src/format_element/document.rs
… crates/rome_formatter/src/format_element/signal.rs
… crates/rome_formatter/src/format_element/tag.rs
… crates/rome_formatter/src/format_extensions.rs
… crates/rome_formatter/src/intersperse.rs
… crates/rome_formatter/src/macros.rs
… crates/rome_formatter/src/printer/call_stack.rs
… crates/rome_formatter/src/printer/line_suffixes.rs
… crates/rome_formatter/src/printer/mod.rs
… crates/rome_formatter/src/printer/queue.rs
… crates/rome_formatter/src/printer/stack.rs
… crates/rome_formatter/src/trivia.rs
… crates/rome_formatter/src/verbatim.rs
… crates/rome_js_formatter/src/builders.rs
… crates/rome_js_formatter/src/check_reformat.rs
… crates/rome_js_formatter/src/js/expressions/template_element.rs
… crates/rome_js_formatter/src/jsx/lists/child_list.rs
… crates/rome_js_formatter/src/lib.rs
… crates/rome_js_formatter/src/syntax_rewriter.rs
… crates/rome_js_formatter/src/ts/lists/type_member_list.rs
… crates/rome_js_formatter/src/utils/assignment_like.rs
… crates/rome_js_formatter/src/utils/mod.rs
… crates/rome_js_formatter/tests/check_reformat.rs
… crates/rome_js_formatter/tests/prettier_tests.rs
… crates/rome_js_formatter/tests/spec_test.rs
… crates/rome_js_formatter/tests/specs/prettier/js/sloppy-mode/eval-arguments-binding.js.snap
… crates/rome_service/src/file_handlers/javascript.rs
… crates/rome_wasm/Cargo.toml
… crates/rome_wasm/build.rs
… npm/rome/tests/daemon/formatContent.test.mjs