@ematipico 2022-09-22
We use the concept of "signal" inside the analyzer, which has its own semantics and it plays will that infrastructure and the actual tool, the transformer. Here, **it seems** that the concept has a different semantic.

My main concern is the overlap of two concepts that have different meanings but same name. It causes confusion *to me*.

A suggestion, maybe could use "Event" instead  - just a suggestion and it's arbitrary -, we have an example of an external crate where it's used: https://github.com/rome/tools/blob/main/xtask/lintdoc/src/main.rs#L221-L243

They look similar to me.