@MichaReiser 2022-09-22
> A suggestion, maybe could use "Event" instead - just a suggestion and it's arbitrary -, we have an example of an external crate where it's used: [`main`/xtask/lintdoc/src/main.rs#L221-L243](https://github.com/rome/tools/blob/main/xtask/lintdoc/src/main.rs?rgh-link-date=2022-09-22T12%3A48%3A05Z#L221-L243)

I'm conflicted on `Event` because it represents the document structure and not that something happened (I guess signal has the same flaw). But I saw that the example uses the term `Tag` which is similar to what the signals represent. Each signal has a start/end tag (maybe rename it to open and close). 

What do you think of the name `Tag`?