@ematipico 2022-09-22
> What do you think of the name `Tag`? 

Yeah that works for me